Exit stuck

+++ file-transfer-lab/Server/fileServer.py
@@ -47,11 +47,11 @@ while True:
             
             #write to file once it's done receiving
             if b".txt" in payload:
-                #save name of file to create/copy file: and later append
-##                if os.path.exists(payload):
-##                    framedSend(sock, b"Error: File already exists in server", debug)
-##                    sys.exit(1)
-                #open file once done
+                #check if file exists
+                if os.path.exists(payload):
+                    framedSend(sock, b"Error: File already exists in server", debug)
+                    sys.exit(1)
+                #open file if it doesnt exist
                 f = open(payload,"wb")
             
             f.write(payload)
